refactor: rename character classes and update package structure
