+++ core/src/main/java/io/github/com/ranie_borges/thejungle/model/entity/characters/Doctor.java
@@ -1,13 +1,13 @@
-package io.github.com.ranie_borges.thejungle.game.characters;
+package io.github.com.ranie_borges.thejungle.model.entity.characters;
 
-import io.github.com.ranie_borges.thejungle.model.Character;
-import io.github.com.ranie_borges.thejungle.model.Item;
-import io.github.com.ranie_borges.thejungle.model.NonPlayerCharacter;
+import io.github.com.ranie_borges.thejungle.model.entity.Character;
+import io.github.com.ranie_borges.thejungle.model.entity.Item;
+import io.github.com.ranie_borges.thejungle.model.entity.NonPlayerCharacter;
 
 //TODO
-public class Survivor extends Character {
+public class Doctor extends Character {
 
-    protected Survivor(String name) {
+    protected Doctor(String name) {
         super(name);
     }
