+++ core/src/main/java/io/github/com/ranie_borges/thejungle/model/entity/characters/Hunter.java
@@ -1,13 +1,13 @@
-package io.github.com.ranie_borges.thejungle.game.characters;
+package io.github.com.ranie_borges.thejungle.model.entity.characters;
 
-import io.github.com.ranie_borges.thejungle.model.Character;
-import io.github.com.ranie_borges.thejungle.model.Item;
-import io.github.com.ranie_borges.thejungle.model.NonPlayerCharacter;
+import io.github.com.ranie_borges.thejungle.model.entity.Character;
+import io.github.com.ranie_borges.thejungle.model.entity.Item;
+import io.github.com.ranie_borges.thejungle.model.entity.NonPlayerCharacter;
 
 //TODO
-public class Lumberjack extends Character {
+public class Hunter extends Character {
 
-    protected Lumberjack(String name) {
+    protected Hunter(String name) {
         super(name);
     }
